requirements.txt
flask>=3.0.0
ffmpeg-python>=0.2.0
Werkzeug>=3.0.0
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
blinker>=1.6.0
python-dotenv>=1.0.0